+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ## [Unreleased]
 
+## [2.0.0] - 2024-01-10
+
+- Add support to redact `Data` objects
+
 ## [1.1.0] - 2021-03-24
 
 - Override `Struct#pretty_print` to keep redacted fields redacted when
